Fixed the Memory Error in _onSourceModelDataChanged
The problem was, that the bottomRight.column() was sometimes exactly
the highest integer value supported by the system. The only reason the error hitted on
Windows and not on Linux was that I have an 32Bit Windows build an a 64bit Linux
build of pyqt.
Two additional checks now assure that the indexes are valid and the
check is not done via range() to preserve memory.
The strange numbers were by the way emitted by QSortFilterProxyModel