Better QML only casting of python objects
The factory now casts recursive. This means that every python value will
recursively casted to a qml counterpart. This is not always what you
want. If you just pass values to qml and back without changing them in
qml its better to stay with QVariant(PyQtObject).
But the factory assumes at the moment you want to read all data in qml